+++ lib/src/live_audio_room.dart
@@ -105,7 +105,7 @@ class _ZegoUIKitPrebuiltLiveAudioRoomState
 
     ZegoUIKit().getZegoUIKitVersion().then((version) {
       ZegoLoggerService.logInfo(
-        'version: zego_uikit_prebuilt_live_audio_room: 2.8.5; $version',
+        'version: zego_uikit_prebuilt_live_audio_room: 2.9.0; $version',
         tag: 'audio room',
         subTag: 'prebuilt',
       );